splitting up modes and executor into two files
