+++ examples/cpp/modes/executor_with_multiple_modes/src/main.cpp
@@ -5,7 +5,7 @@
 
 #include "rclcpp/rclcpp.hpp"
 
-#include <mode.hpp>
+#include <executor.hpp>
 #include <px4_ros2/components/node_with_mode.hpp>
 
 using MyNodeWithModeExecutor = px4_ros2::NodeWithModeExecutor<ModeExecutorTest, FlightModeTestStart, FlightModeTestSegment, FlightModeTestEnd>;
